operator: linear scaling strategy
