+++ k8s/agent-operator/deploy/crds/agentpools.concord.walmartlabs.com-v1.yml
@@ -22,6 +22,8 @@ spec:
             properties:
               autoScale:
                 type: boolean
+              autoScaleStrategy:
+                type: string
               maxSize:
                 type: integer
               minSize:
@@ -43,9 +45,9 @@ spec:
               decrementThresholdFactor:
                 type: number
               queueSelector:
-                additionalProperties:
-                  type: object
                 type: object
+              sizeIncrement:
+                type: number
               configMap:
                 properties:
                   apiVersion:
