+++ k8s/agent-operator/deploy/crds/crd.yml
@@ -2,6 +2,7 @@ apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
   name: agentpools.concord.walmartlabs.com
+  namespace: default
 spec:
   group: concord.walmartlabs.com
   names:
